Hook up the main form layout with child form layouts
Summary: Previously we were only having the wallpaper plugins' child form layouts hooked up to the main form layouts. That causes issues when a label in a plugin is longer than the labels in the main layout. To solve the issue, this patch also makes the main layout responsive to the content of the child layouts.

Test Plan:
Before:
{F6756852}

After:
{F6756855}

Reviewers: #plasma, ngraham, mart

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D20313