Add alternatives button to applet configuration in panel edit mode
Summary:
There is a configure button and an alternatives button in the context menu of an applet, but there is no alternatives button in the popup/overlay of an applet in the panel edit mode, so this adds that button, to make it more consistent and easier to find.
{F6766044}

BUG: 405082
FIXED-IN: 5.16.0

Test Plan: Hover over an applet in the panel edit mode.

Reviewers: #plasma, #vdg, davidedmundson, broulik, ngraham

Reviewed By: #plasma, #vdg, broulik, ngraham

Subscribers: Codezela, broulik, davidedmundson, ngraham, ndavis, #vdg, plasma-devel, #plasma

Tags: #plasma, #vdg

Maniphest Tasks: T10047

Differential Revision: https://phabricator.kde.org/D17410