Allow 'Exclude folders' section in Baloo KCM to fill window
Summary:
Initially, there was a spacer which took up a significant amount of space at the bottom of the window without any apprent justification.

I removed a spacer to allow the folder list to occupy the available space.

It seems like it was added intentionally, so I'm not sure if removing it breaks something. I couldn't find any broken behaviour with this change.

BUG: 407709

Test Plan:
1. Open kcm_baloofile. Check if the exclude folders section fills all available space.
2. Resize the window to make it larger. Check that the exclude folders section resizes to match the window.
3. Open the KCM standalone to and try to resize it smaller than the default size to check if the sizehint is set properly.

Reviewers: #baloo, #vdg, ngraham

Subscribers: GB_2, abetts, bruns, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21414