Filter KNewStuff Plasma Themes by Download Tag
Summary:
Following a lengthy discussion on the topic[1], the first step
towards getting themes cleaned up and hidden when they don't
work is to get the ones hidden which are only for Plasma 4.
We do this by filtering out any theme which is tagged as
being for the major version 4 of plasma. We also allow themes
through which have been tested to work on both, by accepting
themes which are tagged as both 4 and 5.

[1] https://phabricator.kde.org/T8126#184198 and down a few pages

Test Plan:
See themes tagged as plasma##majorversion=4 being hidden, except when they
are also tagged as plasma##majorversion=5. This can most easily be done
by launching the plasma theme dialog and opening Get New Themes from there
and turning on debugging for KNewStuff(Core) like so:

`QT_LOGGING_RULES="org.kde.knewstuff*=true" kcmshell5 kcm_desktoptheme`

and then watching the output, and seeing that content is being rejected
by the InequalityFilter for plasma##majorversion when appropriate.

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21510