Make the Trashcan applet use the same shadow settings as desktop icons
Summary:
We changed desktop shadows in 10278e79f116, but we also need to reflect this change in the Trashcan applet which renders its own label and shadows.

Settings have been copied. This also made it obvious that the label will be blury with non-integer placement so a workaround for that is also provided.

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21545