Move KCMs in "Workspace Theme" to more appropriate locations
Summary:
This is patch one out of three in support of the new {nav Appearance} section layout in Plasma 5.16 decided upon on in {T8871}.

In this patch, KCMs are moved out of the {nav Workspace Theme} category, and the ordering of some KCMs is changed accordingly. The now-empty category gets deleted in D18417.

Test Plan:
{F7134495}

{F7134503}

Reviewers: #vdg, #plasma, davidedmundson, filipf

Reviewed By: #vdg, filipf

Subscribers: filipf, GB_2, abetts, davidedmundson, plasma-devel

Tags: #plasma

Maniphest Tasks: T8871

Differential Revision: https://phabricator.kde.org/D18419