Rename "Look and Feel" KCM to "Global Theme"
Summary:
This implements one of the changes decided upon in T11314. The rationale is as follows:

The old text ("Look and Feel") is a generic term and could be referring to any of the
other appearance settings. The new text "Global Theme" is more descriptive and accurate
and makes this KCM's relationship to everything else clear, especially once it's the only
KCM that uses the word "Theme" (see D22880).

Test Plan: {F7142701}

Reviewers: #vdg, #plasma, GB_2, mart

Reviewed By: #vdg, #plasma, GB_2, mart

Subscribers: mart, GB_2, plasma-devel

Tags: #plasma

Maniphest Tasks: T11314

Differential Revision: https://phabricator.kde.org/D22879