Rename "Plasma Theme" to "Plasma Style"
Summary:
This implements one of the changes decided upon in T11314. The rationale is as follows:

We only use the word "Theme" for LNF settings in the appearance section. For everything else,
we use the word "style" if we have to use anything, preferring to not use any word at all. Since
the app settings use "Style", we should do the same thing for the contrasting Plasma setting.

Test Plan: {F7134998}

Reviewers: #vdg, #plasma, GB_2

Reviewed By: #vdg, GB_2

Subscribers: leinir, GB_2, plasma-devel, kde-doc-english

Tags: #plasma, #documentation

Maniphest Tasks: T11314

Differential Revision: https://phabricator.kde.org/D22880