[Night Color KCM] Rename "Manual" mode to "Constant"
Summary:
Both the Night Color (KDE Plasma) and the Night Light (GNOME Shell) have
so called "Manual" mode. In each case, this mode denotes something
different.

If the Night Color operates in the Manual mode, then the color temperature
remains constant throughout the day. On other hand, the Manual mode in the
Night Light is similar to the Timings mode in the Night Color.

This change renames the Manual mode in order to alleviate the mismatch in
terminology.

Test Plan:
{F7158907}
{F7158912}

Reviewers: #plasma, #vdg, GB_2, ngraham

Reviewed By: #vdg, GB_2, ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22923