Turn the config into the single source of truth
Summary:
The module had a somewhat convoluted logic around the management of its
settings. Sometimes it was reading through Plasma::Theme (mostly
proxying in some form what was in the config), sometimes straight from
KConfig. We now got rid of m_defaultTheme to avoid the temptation of
reading from it, KConfig is our only source of settings.

We kept a single call to Plasma::Theme on save, but that's simply to
ensure proper signalling of the change to the running shell.

Reviewers: #plasma, mart

Reviewed By: #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24392