Integrate the normal keyboard layout switcher
Summary:
This deprecates the SNI from KWin in favour of this plasmoid that will
allow doing the same as well as managing the different IME tech we
support.

This also changes the default panel layout to include kimpanel since it
will be useful in any language.

Related to D12069 and D14796

Test Plan:
I can switch between catalan and american keyboard, if ibus
is running I can do the same but through ibus.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24503