Require and use the new NewStuffQuick Button component
Summary:
This removes the need for the custom hack for spawning a dialog,
and is intended to function as a sample for how to implement this
in other KCMs.

Also use the big, pretty Preview mode (because super-visual content)

This could be done wholesale for the entire bunch of KCMs, but i post
this on its own in an attempt to gain feedback on a self contained
example of how to do this porting step.

Without this patch:
{F7668236}

With this patch:
{F7668234}

Depends on D25041

Reviewers: #plasma, #knewstuff, #vdg, broulik, ngraham

Reviewed By: #vdg, ngraham

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24942