[kcms/krdb] Fix krdb executable
Summary:
krdb is mostly a library used by some KCMs.

It also exists as an executable script that works by invoking kcminit
style which in turn runs the krdb code.

This broke after a port that involved a rename.

Test Plan:
krdb on a distro that doesn't have old files
It ran instead of giving a warning

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25081