kcm/fonts: Do not load image preview before loading settings
Summary:
The way image preview works is it tries to actually "set" and then
"reset" the original value to render the text.

However if we don't have default or user set settings loaded, it will
reset it back to "NotSet" in end instead of user set value and this is
not what we want.

Test Plan:
this alone is not enough to make kcm save subpixel properly
but this is part of patch series

Reviewers: davidedmundson

Reviewed By: davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25090