Notify about change after cursor config is saved
Summary:
Before that patch the cursor change signal was only transmitted after
a configuration was applied on a fly. Because the cursor change is
not applied on a fly on Wayland session yet, the signal is not emitted and
therefore users of that signal are not notified about cursor change.

Now the signal is emitted after the configuration save, so its users are
notified.

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24701