kcm/fonts: set initial settings correctly
Summary:
During initial load we check if certain settings are set or not, and if
they are not set, we set it to rgb/slight (recommended settings by
Plasma team).

however state management of this is broken, previous flow was,

- try to get configuration
- if configuration is not available, set state to desired settings
- set original state to current state value(!)
- write configuration to the file once complete state is written (all
  other settings)

This is fine, when setting is changing from something other then none or
not-set. But if it is not-set, both original state and current state
ended up being desired settings, which write function would happily
ignore since it thinks config never changed.

So, if we encounter not-set configuration, preserve that as a original
setting to get configs written correctly.

This patch on it's own have no visible effect, kcm is still broken due
to lack of kcminit which writes correct configuration at startup, but
this is one of obvious bugfix for follow-up patch series.

Test Plan: with upcoming patch to add kcminit, it correctly applies rgb/slight

Reviewers: #plasma, fvogt, davidedmundson

Reviewed By: fvogt

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25140