[kimpanel] Add place holder button if everything is hide.
Currently, if there is only one button from input method, it is
sometimes impossible to hide any button, because the menu popup may
cause lost of focus and button changes as soon as you right click the
menu. The original logic is to prevent hide all the buttons, but since
we now have placeholder, we can just simply hide any button.