[Workspace KCM] Expose setting to choose what left-clicking in the scrollbar track does
Summary:
This patch exposes a UI for choosing what happens when you left-click in the scrollbar
track  in a Qt or KDE app. The default behavior is unchanged (left-click to page up/page
down) but now you can choose to warp the scroll handle straight to the clicked position.

Having this setting in the Workspace KCM opens the door to later making it also change
the behavior for GTK apps as well, allowing us to remove one more setting from the GTK
KCM and bringing us one step closer to finishing {T10611}.

BUG: 379498
FIXED-IN: 5.18.0

Depends on D25360

Test Plan:
{F7769005}

- Apply D25360
- Click on "Scrolls to the clicked location" in the KCM
- Click in the scrollbar track in a Qt/KDE app
- See that the scroll handle position jumps right there
- Click on  "Scrolls one page up or down" in the KCM
- Click in the scrollbar track in a Qt/KDE app
- See that the scroll handle moves up or down by one page height

Reviewers: #plasma, #vdg, ndavis

Reviewed By: #vdg, ndavis

Subscribers: GB_2, davidedmundson, ndavis, gikari, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25361