[Folder view] Use selected text color for selected items in full representation
Summary:
Right now, selected (not highlighted) items in the popup persist in using
the default text color , which can result in readability problems for light
themes that use a dark highlight effect.

BUG: 414287
FIXED-IN: 5.17.4

Test Plan:
Put a Folder widget in a horizontal panel, click it, hover the mouse over one
of the items in the list, and then click.

Before: {F7772272}

After: {F7772271}

No change or regressions for selected items in Folder View on the desktop.

Reviewers: #vdg, #plasma, ndavis

Reviewed By: #vdg, ndavis

Subscribers: ndavis, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25418