[Plasma Style KCM] Add a color-wheel sign denoting the theme supports the system color scheme
Summary:
The discoverability of the system color scheme awareness feature of some Plasma themes right now is very poor.

In addition to that the previews for Breeze and Breeze Light look identical when using the default color scheme.

To fix this we add a little color wheel icon in the top right corner of the theme preview that only shows up if the theme supports this.

Credits go to @broulik for most of the code, I just positioned the icon.

BUG: 364953

Test Plan:
Depends on: D25999

{F7818015}

Doesn't get messed up with scaling either.

Reviewers: #plasma, #vdg, ngraham

Reviewed By: #vdg, ngraham

Subscribers: ngraham, broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25785