Add grouping and "Show Only Minimized" options to icons-only task manager
Summary:
FEATURE: 368430
FIXED-IN: 5.18.0
Being able to disable grouping for the icons-only task manager makes sense, because it has enough space to hold plenty of application icons/launchers. The "Show only tasks that are minimized" option is useful as well for the IOTM.
{F7258179}

Test Plan: Open the IOTM behavior settings page. Enable/disable grouping and toggle the "Show only tasks that are minimized" option.

Reviewers: #plasma, #vdg, ngraham

Reviewed By: #vdg, ngraham

Subscribers: ngraham, hein, #vdg, plasma-devel, #plasma

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23242