[WIP] Switch the Attica KDE plugin to use KAccounts
Summary:
Don't require KWallet, and require KAccounts and libaccounts-Qt5

Link and use includes from new requirements

Remove KWallet bits, and use KAccounts/libaccounts-qt instead

This removes the custom KWallet integration in favour of an approach
based on the more generic KAccounts framework. It is a work in progress
which currently does not allow for creating new accounts (which would
be done through saveCredentials), but which does load and use the data
from any accounts which already exist.

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25961