[Plasma Style KCM] Add search filter
Summary:
Addsm the ability to filter themes by name by typing in the search line or
whether they use the current color scheme or if they are light or dark if they
do not. To enable this the model is extracted out into its own class and a
QSortFilterProxyModel is added for the Filtering. This is mostly based on the
same functionality of the Colors KCM and uses the same heuristic to decide if a
theme is light or dark.

Test Plan: {F7821003}

Reviewers: #plasma, #vdg, broulik, ndavis, ngraham, ervin

Reviewed By: #vdg, ndavis, ngraham, ervin

Subscribers: ervin, ndavis, crossi, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26039