[Application Style] Add GTK Application Style Page
Summary:
The GTK KCM is now moved to Application Style KCM as a sub page. The
functionality is identical to the one, found in GTK KCM, as well for the
bugs :)

The page is only accessible, if the `gtkconfig` kded module is loaded,
because only with the module the configuration can be changed.

Depends on D26261

Test Plan:
# Apply D26261
# **Restart kded5**
# Open System Settings, go to Application Style KCM
# The button for GTK Configuration should be present
# Go to Startup and Shutdown > Background Services, disable `Plasma GTKd` {F7877157}
# Return to the App Style KCM, the button should disappear
# Return to the Kded KCM and enable `Plasma GTKd` back
# Return to the App Style KCM - button is present now
# Go to subpage
# Happy (I hope) testing!

{F7877153}

{F7877154}

Reviewers: #vdg, broulik, #plasma, ngraham, mart

Reviewed By: #vdg, #plasma, ngraham, mart

Subscribers: davidedmundson, ngraham, plasma-devel

Tags: #plasma

Maniphest Tasks: T10611

Differential Revision: https://phabricator.kde.org/D26583