Make the lookandfeel KCM build without XCursor support
Summary:
With this change the lookandfeel KCM can be built without X11 cursor development headers.
In the case the KCM is built without XCursor support, it will not be able to reload the cursor dynamically at runtime but it will still be able to configure cursor themes.

This is an alternative to: D26344 (https://phabricator.kde.org/D26344)

Test Plan: plasma-desktop builds using kdesrc-build with and without X11 cursor development headers present.

Reviewers: davidedmundson, #plasma

Reviewed By: davidedmundson, #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26367