DesktopPaths KCM: Move the settings logic to a KCoreConfigSkeleton class
Summary:
Our KCoreConfigSkeleton subclass is interestingly hand written since
it's too remote from what we usually do (the resulting file needs to be
sourceable and we don't want the usual behavior of removing keys when
they're set to defaults).

The only one which was "more standard" was the autostart path, but
really with the URL conversion on top and for just a single entry, I
shoved it in the hand written class as well.

Doesn't reduce the code much, but at least it separates that logic from
the KCM and it opens the door to properly working defaults button.

Reviewers: #plasma, crossi, bport, meven, mart, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26389