DesktopPaths KCM: Remove the moving directory logic
Summary:
This simplifies greatly this otherwise mundane KCM. It introduced a lot
of complexity while gettings things wrong anyway: it would miss if a
directory is included in another one except in the "autostart inside
desktop" case, it would also leave you stranded in case of an
uncompleted move.

So now the KCM has a clear purpose: edit the settings. Moving data
around is still up to the user using a proper filemanager.

Long term it would probably be better for the filemanager to edit the
settings automatically when the user move on of those folders around,
then the KCM could be decommissioned.

Reviewers: #plasma, crossi, bport, meven, mart, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26391