[Colors KCM] Correctly report whether saving is needed
Summary:
Previously the kcm always reported that a save is needed once a scheme had been
marked for deletion. The correct way is to check if there are schemes that are
currently marked for deletion. Also we don't need the firstTime workaround.

Test Plan: Mark a scheme for deletion and undo that action

Reviewers: #plasma, mart

Reviewed By: #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26526