[Notifications KCM] Force re-evaluation of position radio buttons
The Settings object is kind of a write-only state and doesn't signal property changes unless "live" (where it would listen to other
settings changing which we don't want inside the KCM) or only upon save().
This is arguably a questionable design decision on my side :-) but usually it's not a problem as the KCM controls will have
whatever state the user will put them in but it falls apart in case we switch to another page and back without saving.

BUG: 414098
FIXED-IN: 5.18.0

Differential Revision: https://phabricator.kde.org/D26162