[KDED KCM] Rewrite as KDeclarative ScrollViewKCM
This rewrites the "Background services" KCM in QML using ScrollViewKCM.
The two separate list views are merged into a single one with the configurable services at the top,
and the ones that are loaded on-demand and "only for your convenience" at the bottom.
A search field is added searching through name and plugin ID. Since the sortable table headers are gone,
a filter combo is provided instead to filter for all, running, or non-running services.
As an extra Schmankerl when starting a service that immediately disables itself again (which technically isn't an error
that would be indicated as such) a hint is shown you're not left wondering why it doesn't start.
A hint is also displayed when services got automatically started/stopped when applying settings as this reloads kded5.
Furthermore, the code is cleaned up a lot (quite eerie, adding a 2020 Copyright to an existing 2002 one :),
ported to json plugin data, and a proper QAbstractListModel added.

Differential Revision: https://phabricator.kde.org/D26506