[widgetexplorer] Fix animate in animation
Summary:
The animation in comes in from offscreen on the left.

"to: 0" was correct when we had one column. With the 2 column view it
means we animate two items on top of each other and jump when the
animation completes. We want to animate to the final position the item
will end up in.

It bugs me more than it probably should.

Test Plan:
Search for "asdfasdfsadfsadf"
Remove search
Animation now looks ok, instead of horrific

Reviewers: #plasma, ngraham

Reviewed By: ngraham

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26698