[Folder View] Fix text label color in pop-up
Summary:
Currently there's a top-level ColorScope that is almost entirely unused except for
introducing a bug whereby the label text in the pop-up is the wrong color. Removing the
ColorScope fixes the bug and does not change the appearance of anything else anywhere else.

BUG: 416658
FIXED-IN: 5.18.0

Test Plan:
Hover over a folder and click on button to open a folder pop-up
See that the text color is correct now: {F7925905, size=full}

Use Folder view as a containment, as a widget, and in a panel
See that text color is fine everywhere

Use a bunch of different color schemes and repeat the above tests and see that everything
still looks good

Reviewers: #vdg, #plasma, hein, davidedmundson, mart

Reviewed By: #vdg, #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26884