Add kconf update script for ksmserver
Summary:
A config enum value was moved in 506d1807e6a444b6f0fc252b21e8b16b1cf1b2e4 to
kcfg we should port the value for existing setups.

What used to be the string "default" is now "emptySession"
"default" was not the default. (confusingly)

On an upgrade "default" will fail to load and go to the default which is "restorePreviousLogout" and not "emptySession"

Test Plan: Ran script

Reviewers: #plasma

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25634