Fix drag and drop from taskbar to activity pager
The code to handle a window drag from the taskbar to drop on the
Activity Pager to add the window to an Activity was broken,
because the test to see if the drop location is a running activity
still treated the itemId as an integer indexing into an array of running
activity indices, rather than an id string that could be a member of the
list of running activities.

It is not clear whether this is a bug fix or
a new feature.  The drop handler in PagerModel.cpp I amended is not used
when DnDing windows within the pager, but may have been used for that at
one point.

This also allows the familiar file manager semantics of ctrl+drag = copy,
drag = move to be used when dragging and dropping windows between
activities using the taskbar and the activity pager applet.