Set toolbox to invisble when it's hidden
Summary:
opacity was bound to editMode, but this did not affect visibility or
enabled.

Not only is this somewhat wasteful, enabled is very important as
otherwise shortcuts for buttons get intercepted to match the accelerator
applied to the internal buttons. This breaks all the multi-key shortcuts
that plasma uses as alt+d gets taken.

BUG: 416695
FIXED-IN: 5.18.1

Test Plan:
Mad gammaray skillz to find the bug

Checked alt+d, a now opens the widget explorer
Checked edit mode through the context menu still showed the toolbox

Reviewers: #plasma, ngraham

Reviewed By: ngraham

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27386