[desktoppackage] Fix layout when an applet has no custom config modules
Summary:
There was some broken code to hide the sidebar when we only have one
module. However at some point it seems half was removed, half wasn't.

This leads to the content overlapping the sidebar frame, but the sidebar
was invisible and it looks generally broken.

This patch completes the removal.

BUG: 410770

Test Plan:
Opened touchpad KCM
Saw the one module (keyboard shortcuts)
Sidebar was visible, and nothing was overlapping

Reviewers: ngraham

Reviewed By: ngraham

Subscribers: broulik, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27398