Calculate needsSave on checkbox toggle not on change
Summary:
Otherwise we will trigger it during load, which is wasteful.

But worse potentially we haven't evaluated the other settings yet, so it
could trigger that we have changed the settings, and prematurely set
apply to enabled prematurely which currently breaks things.

I suspect it will fix the bug below, I couldn't reproduce reliably
enough to prove it.

Please reopen if it remains an issue

BUG: 411584
Fixed-in: 5.18.2

Test Plan:
Opened KCM
toggling still worked as expected

Reviewers: #plasma, apol, ngraham

Reviewed By: apol, ngraham

Subscribers: apol, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27590