[emojier] Set minimum window height and width
Summary:
There are two problems right now with the emojier's window:
- when shrinking the window a scrollbar is drawn *over* the category drawer's icons
- the whole window can be shrunk to be only 1px wide and only several pixels tall

To solve both issues this patch sets a minimum height that is equal to the height of the category drawer's content, thereby removing the possibility of a scrollbar appearing. Minimum width is then set to be somewhat bigger than minimum height to account for monitor aspect ratios.

In practice this amounts to about 465x370 on a 1080p screen.

BUG: 418195
FIXED-IN: 5.18.3

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27679