Solid-device-automounter/kcm: show disconnected known device when disconnecting it
Summary:
A user might find odd that when a known device in the UI is unmounted, it disappears.
Despite the fact that restarting the kcm would show it.

Test Plan:
Open `kcmshell5 device_automounter_kcm`
Plug in a usb stick, set a setting for it and Apply
Unplug the stick

Before:
The disconnected usb stick does not appear
After:
The usb stick appears at the bottom of the disconnected section

Forget the stick in the UI and Apply
Plug in the usb stick
Unplug it

The usb stick is not visible (as before)

Reviewers: broulik, #plasma, ervin, crossi, bport

Reviewed By: ervin, bport

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27024