[Baloo KCM] Complete overhaul of the include/exclude folder list
Summary:
The current "Excluded folders" list in the KCM is quite awkward:

1. It tries to mimic baloos automatic exclusion of external drives,
   but fails doing so and adds almost any external drive even when not
   mounted below an indexable path.

2. Deleting an autogenerated entry actually **adds** it to the included
   folder list, and then hides it.

3. There is no way to show the included folder list, or add any entries
   to it.

Remove the custom "excluded mounts" heuristic from the KCM and retrieve
the additional (not explicitly configured) excluded ones from baloo.

Replace the "excluded list" with a common list for included and excluded
folders, and flag its state. This also makes it easy to add additional
properties later.

Create a new UI delegate for each config list item, allowing to enable
and disable indexing for each entry. Make the "delete" actually always
delete a config entry, and make the control inline. Move the "Add" button
to the *right* bottom of the list (in accordance with UI guidelines) and
add some text to it.

Depends on D28024

Solves the following parts of T9879:
[x] Allow adding of includeFolders in the KCM
[x] Correctly show excludeFolders below explicit (non-$HOME) includeFolders in KCM

CCBUG: 417170
CCBUG: 366521
BUG: 417763
BUG: 417762

Reviewers: #baloo, #vdg, ngraham, mart

Reviewed By: #baloo, #vdg, ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28025