[Icons KCM] Floor delegate height
There's two rows, so if the delegate happens to be at an odd size, both rows would be rounded up,
pushing the icons too close to the delegate border.
Since there's already code for unloading the icons that are out of view, nothing should be leaked into the view,
even if we now end up having 1 or 2 pixels of the delegates normally out of view visible.

Differential Revision: https://phabricator.kde.org/D28120