Add the option to show the current activity name and icon
Summary:
The current activity switching applet shows a generic icon that
opens the activity switcher which does not offer any information
on the current activity. This does not go well with other applets
like desktop pager.

This patch adds options (on by default) to show the current activity
icon and name in the applet.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28136