[kcms/baloo] Refine UI and prepare for more indexing options in the future
Summary:
This patch refines the UI added in D28025. The basic goal here is to to be able to choose
what type of indexing is done for each path (none, full, or metadata only). I wrestled
for a while with the best way to present this information and also allow it to be easily
changed and eventually settled on a combobox, given that it combines both functions into
one compact control.

The code is made flexible so that once metadata-only indexing is added in the future, it
will be easy to add the appropriate option into the UI.

Test Plan: {F8198344}

Reviewers: #vdg, #baloo, #plasma, ndavis

Reviewed By: #vdg, ndavis

Subscribers: ndavis, bruns, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28280