KCM/Mouse: fix the middleEmulation checkbox status
Summary:
The load function of the field was not called previously.
Preventing the checkbox from reflecting the property status and if it is enabled.

Test Plan:
kcmshell5 mouse
Select a pointer not support middle click emulation.
Before:
middle click emulation is enabled
After:
middle click emulation is disabled

Reviewers: #plasma, ervin, bport, crossi, hchain

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28296