Make accidental panel and panel widget deletion more difficult
Summary:
We make it very easy to delete panels and panel widgets--perhaps a bit too easy. User
feedback has indicated that we may need to re-work this a bit, and that perhaps D20144 was a
mistake.

This patch moves the {nav Remove Panel} button back into the {nav More} menu where it used to
be, to make it harder to accidentally click on. It also moves the {nav Remove} button for
panel widgets higher up in the menu so it's not so close to the cursor.

To adjust to these changes, the layout of buttons on the panel toolbar is changed to
preserve visual balance, and the {nav More} menu's name is adjusted to reflect the fact that
removing a panel is not a setting.

BUG: 419853
FIXED-IN: 5.18.5

Test Plan: {F8227561}

Reviewers: #vdg, #plasma, ndavis

Reviewed By: #vdg, ndavis

Subscribers: rikmills, broulik, plasma-devel

Tags: #plasma

Maniphest Tasks: T12942

Differential Revision: https://phabricator.kde.org/D28710