[Fonts KCM] Remove redundant nearestExistingFont()
Summary:
It seems that nearestExistingFont() gets the same result as
`fc-match font-name`, which is basically the sans serif default font
on the system; this seems redundant as KFontChooser will fallback to
selecting the first font family in the list if the initial font it
was called with doesn't exist _and_ KFontChooser always puts "Sans Serif",
"Serif" and "Monospace" at the top of the list.

Removing nearestExistingFont() solves the issue in bug 420287 as
the font it returns will have the styleName property set, after we
went to so much trouble to clear that property so that setBold() can
work.

CCBUG: 420287

Test Plan:
Before the patch:
- In kdeglobals [General] set:
  fixed=Blah Mono,12,-1,5,50,0,0,0,0,0
  toolBarFont=Bogus Serif,12,-1,5,50,0,0,0,0,0
- Load the fonts KCM, and open the font dialog for Fixed and Toolbar,
  in both cases the default "sans serif" font on the system is selected,
  in my case it's "DejaVu Sans"

Apply the patch and repeat, the "Sans Serif" entry is selected, which is
an alias to "DejaVu Sans" on my system.

Reviewers: #plasma, bport, ngraham

Reviewed By: bport, ngraham

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29155